bugfix: added a monitor to control a race condition in the dump module
Rationale:
In multi-thread, the creation of a new output directory is held but
a thread different than the main one. A race condition appears
if the main thread wants to write a packet to a trace but the other
thread commands the creation of a new directory so it closes all
the old traces.


git-svn-id: http://tstat.polito.it/svn/software/tstat/trunk@211 da00f6ee-164e-4028-97e9-5fc1dd4040cf